+++ src/main/java/fun/gen/GenCache.java
@@ -2,7 +2,6 @@ package fun.gen;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Objects;
 
 import static java.util.Objects.requireNonNull;