+++ docs/CHANGELOG.md
@@ -1,3 +1,8 @@
 1.4.0
   New feature:
-    - Named generators:
+    - Named generators: NamedGen 
+  Refactor: 
+    - JsObjGen (improvement nullable and optional fields generation)
+
+  Doc:
+    - New section in readme about recursive generators